feat(refactor): 클래스 정리
**
Data Enum들을 enums패키지에 통합
**
