+++ src/main/java/com/sprint/mission/discodeit/entity/Channel.java
@@ -1,5 +1,7 @@
 package com.sprint.mission.discodeit.entity;
 
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
+
 import java.util.UUID;
 
 /**
