+++ src/main/java/com/sprint/mission/discodeit/entity/enums/ChannelType.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.entity.enums;
 
 public enum ChannelType {
     CHAT("대화"),
