+++ src/main/java/com/sprint/mission/discodeit/entity/enums/Role.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.entity.enums;
 
 public enum Role {
     ADMIN("관리자"),
