+++ src/main/java/com/sprint/mission/discodeit/repository/ChannelRepository.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.repository;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
 
 import java.util.List;
 import java.util.Optional;
