+++ src/main/java/com/sprint/mission/discodeit/repository/impl/ChannelRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.repository.impl;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import org.springframework.stereotype.Repository;
