+++ src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
 
 import java.util.List;
 import java.util.UUID;
