+++ src/main/java/com/sprint/mission/discodeit/service/impl/ChannelServiceImpl.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service.impl;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import org.springframework.stereotype.Service;
