+++ src/main/java/com/sprint/mission/discodeit/service/impl/ParticipationServiceImpl.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service.impl;
 
 import com.sprint.mission.discodeit.entity.Participation;
-import com.sprint.mission.discodeit.entity.Role;
+import com.sprint.mission.discodeit.entity.enums.Role;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.ParticipationRepository;
