+++ src/main/java/com/sprint/mission/discodeit/view/ChannelView.java
@@ -1,10 +1,8 @@
 package com.sprint.mission.discodeit.view;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
 import com.sprint.mission.discodeit.service.ChannelService;
-import com.sprint.mission.discodeit.service.impl.ChannelServiceImpl;
-import com.sprint.mission.discodeit.utils.AppConfigRegacy;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
