+++ src/test/java/com/sprint/mission/discodeit/JCFParticipationServiceTest.java
@@ -2,7 +2,7 @@ package com.sprint.mission.discodeit;
 
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Participation;
-import com.sprint.mission.discodeit.entity.Role;
+import com.sprint.mission.discodeit.entity.enums.Role;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.impl.ChannelRepositoryImpl;
 import com.sprint.mission.discodeit.repository.impl.ParticipationRepositoryImpl;
